Rename `.github/README.md` file to `ABOUT.md`
Currently, when opening the OSCAL repo on GitHub, the `README.md` that is shown is take from `./.github/README.md` instead of `./README.md`.
This is a documented behavior of GitHub which can be seen [here](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-readmes).
To make the repo better documented, I renamed the `.github/README.md` file to `ABOUT.md` file so the real (and main) `README.md` file would be shown when opening the repo on GitHub.