Change GitHub links to point to `main` branch instead of `master`
There are currently some links which point to GitHub repos with the `master` branch.
All the repos they're pointing to don't have a `master` branch anymore since it was replaced with `main`.
This PR fixes the links.

FYI - The links still currently work, but show an annoying "Branch master was renamed to main." message every time you follow them.