Boolean handling workaround for workflow_call vs workflow_dispatch bug. (#1236)
This fix will resolve the issue noticed in the runs post release prep.

https://github.com/usnistgov/OSCAL/runs/6311164174
https://github.com/usnistgov/OSCAL/runs/6311163976

This is a workaround report in this thread.

https://github.com/github/feedback/discussions/9343#discussion-3768151

It pertains to this issue, which has yet to be fixed in GitHub Actions
runners that exhibit a difference in behavior for the same boolean input
being treated as a string in a `workflow_dispatch` `inputs`
(i.e. `github.event.inputs.*`) even if `type: boolean` is set but
behaves properly as intended for `workflow_call` `inputs`
(`inputs.*` which are different in the object graph) and treated as bools
through and through. This is reported in actions/runner#1483.

To resolve this, the buggy `workflow_dispatch` `github.event.inputs.*`
will be left with the workaround as-is, conditional checking the values
as strings, but we revert the changes for `workflow_call` `inputs.*` and
leave them like before until actions/runner#1483 is resolved.