submodule update fixes model doc bug usnistgov/metaschema#225.
It appears whatever was reported and was happening reliably when I was
using the current commit for usnistgov/OSCAL develop branch module[0]
has now gone away when updating a few commits to current develop commit
in the submodule[1], ahead of usnistgov/OSCAL check-in. It appears it
was maybe fixed between the following commit range.

[0] https://github.com/usnistgov/metaschema/commit/f28a588f0e9ac013125ed0f118f0fcaef4d1a6e2
[1] https://github.com/usnistgov/metaschema/commit/a11515b2d431abb031921ca6c73b83d007b229a7

Can attempt to do `git blame` or further analysis at a later date.