Update metaschema submodule for usnistgov/OSCAL#1454. (#1455)
* Update metaschema submodule for usnistgov/OSCAL#1454.

* Update obsolete GCN FR OSCAL article link

While working on usnistgov/OSCAL#1454, this link appears to have changed
on the GCN news site, so we update. More info in failed build here.

https://github.com/usnistgov/OSCAL/actions/runs/3054991503

Additionaly we formatted it back down to a conventional Markdown link.
Previously it was an inline <a/> with data-proofer-ignore='yes' attribute
and it was/is unclear why that was done. No conclusive evidence found
with git blamen on a call with Dave.

Be forewarned: if you found this commit to understand, you have likely
gone too far.