Fix documentation relative paths in Metaschema
Recently added links in documentation refer to href="/OSCAL" and not
href="https://pages.nist.gov/OSCAL" in error or to simplify link gen
and lookup. Now that we are remapping the full URLs to local paths and
offsets for link-checking, we want these to be consistent across the
board in Metaschema after usnistgov/OSCAL#1263.