[WIP] Test manually adding transforms.
It appears that I had not completely understood the flow of CI, then CD,
in the project. After updating the submodule, it should have been
obvious to me that the updated converters are not committed statically,
but dynamically based on what branch it is. If it is a PR branch, the
generated converters are not updated or committed. For now, we will Test
updating them manually, and then override the commit.