Update Actions to Properly Use PAT for Website Build and Deployment (#1726) (#1731)
* Correct GHA action-gh-pages argument for #1726

By doing this, we correctly the PAT usage and not ironically use an existing,
but improperly permissioned GITHUB_TOKEN provided as a context machine identity
for all runs of all workflows, this should fix the builds and stop the cryptic
HTTP 429 rate limit error response. It's cryptic because you get a 429 response
after one single API operation (with git clone) because the token is wrong.

* Remove lychee scans during site build for in #1726

As part of the troubleshooting work, GH docs do indicate scanning links from the GHA
runners can potentially cause rate limiting. We have automated nightly scans and we
review code changes as part of PRs. We can forgot commit-by-commit link scanning as
a short-to-medium term mitigation and enable it again later.

* Add @Compton-NIST's rate limit checks from #1698 for #1726.