Datatype regex fixes for #1703 (#1736)
* Update metaschema for current datatype regexes for #1703.

* Back-port Metaschema XSD relocation.

As part of the OSCAL build process, we validate the structure of the
metaschema definitions of the OSCAL models themselves against the schema
of Metaschema definitions XSD from the metaschema repo. Recently, a path
update was fixed in the OSCAL develop branch but not in the release-1.0
branch or main because these changes occurred after 1.0.4 publication.
The files were relocated and reorganized. This adjustment must be made
to account for that.

Pull release/main back into develop should be minor and a rebase may not
be necessary. If so, we will adjust accordingly. In develop, see This
commit:

https://github.com/usnistgov/OSCAL/commit/84d2d4659f2db3948160fb92566e56f79f922b89

* Workaround maven.restlet.org service cert expiry. (#1602)

This is pulling over workaround as reported in usnistgov/oscal-content#139 with the WIP workaround submodule.

usnistgov/oscal-content@26f0fe1

* Update metaschema submodule again.