Update metaschema for 1.0.6 release (#1820)
* Update metaschema for 1.0.6 release

This PR commit will update the Metaschema submodule to update. We need
to update the underlying version of the MIMF to address the following:

Closes usnistgov/OSCAL#1773.
Closes usnistgov/OSCAL#1774.

* Fix Metaschema XML schema URL

Per @nikita-wootten-nist's feedback, fix this URL which has not been updated in either
develop or main branches as the Metaschema 0.9.0 and directory restructuring occurred
during the course of the drafting of the 1.0.6 release.