Add back pom.xml for Maven deps
This was removed in site migration work and did not appear to be relocated, as determined during the course of usnistgov/oscal-content#204. We can rely on the metaschema-xslt project, but given OSCAL has the profile resolver tool committed in its repo tree, it stands to reason to rely on its own independent Maven POM file, even if now redundant, as they may diverge in the future.