Do not require a `protocol/@name` for #1772
Per discussion with community members and the nature of port, protocol,
and service declarations in a OSCAL SSP model instances for RMF use
cases, like FedRAMP and others. It would appear the model, per the
Metaschema declarations and documentation, require a port range has a
name that is commonly the IANA service name, which should be optional.
Otherwise, developers and security officials will need to create an
arbitrary name that does not strictly conform to the documentation. More
details can be found in the issue thread referenced below by URL.

https://github.com/usnistgov/OSCAL/issues/1772#issuecomment-2466441626